refactor: move mapper package to API package.
