+++ identity/src/main/kotlin/net/thechance/identity/api/mapper/UserMapper.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.mapper
+package net.thechance.identity.api.mapper
 
 import net.thechance.identity.api.dto.ProfileResponse
 import net.thechance.identity.entity.User
